requirements.txt
fastapi
uvicorn
groq
google-generativeai
python-dotenv
asyncio